initial refactor to controlinterface
